the big coordinate change
In this patch, change the whole Map interface so that methods accept
(const Coordinate&) instead of pair of numbers. This rather invasive
change impact a lot of the code, but the modification is, in most
case, quite simple.

In a lot of case, this allows for simplication of `Map::foo(pos.X(),
pos.Y())` into `Map::foo(pos)`, improving readability. In other
places, one had to instead add more `.X() .Y()` call instead of using
the formerly existing `x, y` private variables, so the code gets
heavier. This uses will for the most part be simplified in the
following commits: this one focus on trivial change, while more
important restructuring are coming later. Note that we already make
good use of some of the new Coordinate or Random functions.