initial blob to use libtcod event system
Either the previous libtcod didn't have an event system, or there were some modifications to the bundled libtcod to make it work without what's currently available.

In either case, add evented stuff (with two FIXME notices for future rework) to the main menu. This allows saving the game, and using the mouse to select stuff (in the main menu only). Other menus are in the works.